Test coulomb scattering

+++ utils/integrate.hpp
@@ -13,6 +13,9 @@
 #include "gen_ex.hpp"
 #include "spline.hpp"
 
+//this is out of date
+//consider using chebyshev instead
+
 class cum_adap_simps_data
 {
 private:
@@ -157,7 +160,7 @@ public:
 		//step two, if necisary, refine the lower two sub sections
 		//if( std::abs( rough_quad() - (left_data->rough_quad()+right_data->rough_quad())) > 15*rel_tol )
 		double R=rough_quad();
-		if( float(R*factor + ( R - (left_data->rough_quad()+right_data->rough_quad()))) != float(R*factor) )
+		if( float(R*factor + float( R - (left_data->rough_quad()+right_data->rough_quad()))) != float(R*factor) )
 		{
 		    //print("RECURE!");
 			left_data->refine(factor);
